folder restructure to follow python best practices
