+++ hexgame/color.py
